+++ CHANGELOG.md
@@ -7,10 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add tests for `VersionConfig.{filter_targets,filter_dates}`
+
 ### Fixes
 
 - Ensure `.build/` directory exists even when there are no targets.
-- Fix `VersionConfig.filter_targets` + add tests
+- Fix `VersionConfig.filter_targets`
 
 ### Dependencies